+++ src/analysis/compiled_terminal.rs
@@ -27,8 +27,8 @@ impl CompiledTerminal {
         R: Fn(&str) -> TerminalIndex,
     {
         match s {
-            Symbol::T(Terminal::Trm(t, _)) => Self(terminal_index_resolver(t)),
-            Symbol::T(Terminal::End) => Self(EOI),
+            Symbol::T(Terminal::Trm(t, _), _) => Self(terminal_index_resolver(t)),
+            Symbol::T(Terminal::End, _) => Self(EOI),
             _ => panic!("Unexpected symbol type: {:?}", s),
         }
     }