+++ src/analysis/first.rs
@@ -182,14 +182,14 @@ fn combine_production_equation<'a, 'c: 'a>(
                 // For each non-terminal create a separate SymbolString
                 Symbol::N(_, _) => acc.push(SymbolString(vec![s.clone()])),
                 // Stack terminals as long as possible
-                Symbol::T(_) => {
+                Symbol::T(_, _) => {
                     if acc.is_empty() {
                         acc.push(SymbolString(vec![s.clone()]));
                     } else {
                         let last = acc.len() - 1;
                         let last_len = acc[last].0.len();
                         let last_terminal = &acc[last].0[last_len - 1];
-                        if matches!(last_terminal, Symbol::T(_)) {
+                        if matches!(last_terminal, Symbol::T(_, _)) {
                             // Only add to terminals
                             acc[last].0.push(s.clone());
                         } else {
@@ -198,6 +198,7 @@ fn combine_production_equation<'a, 'c: 'a>(
                         }
                     }
                 }
+                Symbol::P(_) => (),
                 Symbol::S(_) => (),
                 Symbol::Push(_) => (),
                 Symbol::Pop => (),
@@ -223,7 +224,7 @@ fn combine_production_equation<'a, 'c: 'a>(
     for symbol_string in parts {
         // trace!(" + {}", symbol_string);
         match &symbol_string.0[0] {
-            Symbol::T(_) => {
+            Symbol::T(_, _) => {
                 result_function = Box::new(move |result_vector: &ResultVector| {
                     let mapper = |s| CompiledTerminal::create(s, terminal_index.clone());
                     result_function(result_vector).k_concat(
@@ -238,6 +239,7 @@ fn combine_production_equation<'a, 'c: 'a>(
                     result_function(result_vector).k_concat(&f(result_vector), k)
                 });
             }
+            Symbol::P(_) => (),
             Symbol::S(_) => (),
             Symbol::Push(_) => (),
             Symbol::Pop => (),