+++ src/grammar/cfg.rs
@@ -107,7 +107,7 @@ impl Cfg {
     pub fn get_ordered_terminals(&self) -> Vec<(&str, Vec<usize>)> {
         self.pr.iter().fold(Vec::new(), |mut acc, p| {
             acc = p.get_r().iter().fold(acc, |mut acc, s| {
-                if let Symbol::T(Terminal::Trm(t, s)) = s {
+                if let Symbol::T(Terminal::Trm(t, s), _) = s {
                     if let Some(pos) = acc.iter_mut().position(|(trm, _)| trm == t) {
                         for st in s {
                             if !acc[pos].1.contains(st) {
@@ -134,8 +134,8 @@ impl Cfg {
             .enumerate()
             .fold(BTreeMap::new(), |mut acc, (pi, p)| {
                 acc = p.get_r().iter().enumerate().fold(acc, |mut acc, (si, s)| {
-                    if matches!(s, Symbol::T(Terminal::Trm(_, _)))
-                        || matches!(s, Symbol::T(Terminal::End))
+                    if matches!(s, Symbol::T(Terminal::Trm(_, _), _))
+                        || matches!(s, Symbol::T(Terminal::End, _))
                     {
                         acc.insert(Pos::new(pi, si + 1), s);
                     }