+++ CHANGELOG.md
@@ -21,6 +21,7 @@ But we try to mark incompatible changes with a new minor version.
 
 * Removed serialization support - no use case anymore
 * Removed some useless derives
+* Introduced pseudo symbol
 
 ## v0.8.4-pre - Not released yet